Using NodeUtils only in node
