+++ packages/core/js-peer/src/services/_aqua/services.ts
@@ -6,9 +6,8 @@
  * Aqua version: 0.7.7-362
  *
  */
-import { CallParams } from '@fluencelabs/interfaces';
+import { CallParams, IFluenceInternalApi } from '@fluencelabs/interfaces';
 import { registerService } from '../../compilerSupport/registerService.js';
-import { FluencePeer } from '../../jsPeer/FluencePeer.js';
 
 // Services
 
@@ -27,7 +26,7 @@ export interface SigDef {
     ) => boolean | Promise<boolean>;
 }
 
-export function registerSig(peer: FluencePeer, serviceId: string, service: any) {
+export function registerSig(peer: IFluenceInternalApi, serviceId: string, service: any) {
     registerService({
         peer: peer as any,
         service: service,
